requirements.txt
# Core web framework
Django>=4.2

# Vector database
chromadb>=0.4.22

# PDF processing / OCR
PyMuPDF>=1.23.0
Pillow>=10.0.0

# AI / LLM clients
ollama>=0.1.7
huggingface_hub>=0.20.0
transformers>=4.40.0
accelerate>=0.20.0
sentencepiece>=0.1.99
google-genai

# ML / Numerical
numpy>=1.24.0
torch>=2.0.0

# Django add-ons
django-cors-headers>=4.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.1

# Testing
pytest>=7.0.0